requirements.txt
numpy
pillow
qrcode
fastapi
uvicorn[standard]
requests